+++ classes/include/DwmIpAddress.hh
@@ -137,7 +137,7 @@ namespace Dwm {
     //!  Returns the number of bytes that would be written if we called
     //!  one of the non-compressing Write() members.
     //------------------------------------------------------------------------
-    uint32_t StreamedLength() const;
+    uint64_t StreamedLength() const;
     
     //------------------------------------------------------------------------
     //!  Reads from an istream.  Returns the istream.
