+++ classes/include/DwmIpv4UdpHeader.hh
@@ -207,7 +207,7 @@ namespace Dwm {
     //!  Returns the number of bytes that would be written if we called one
     //!  of the Write() members.
     //------------------------------------------------------------------------
-    uint32_t StreamedLength() const override;
+    uint64_t StreamedLength() const override;
 
     //------------------------------------------------------------------------
     //!  Reads the UDP header from a gzFile.  Returns the number of bytes
